refactor: refactor SearchField component, update styles
