+++ src/components/ui/form/searchField/SearchField.tsx
@@ -6,6 +6,9 @@ import { variantStyles } from '@/constants/borderVariants.constants'
 import { useQuery } from '@tanstack/react-query'
 import { X } from 'lucide-react'
 import { LoaderCircle } from 'lucide-react'
+import { useDebounce } from '@/hooks/useDebounce'
+import { fetchSearchResults } from '@/services/transactionApi/apiSearchISIN'
+import type { SearchResult } from '@/services/transactionApi/apiSearchISIN'
 
 interface SearchFieldProps extends Omit<InputHTMLAttributes<HTMLInputElement>, 'onChange' | 'value' | 'onSelect'> {
     label: string
@@ -13,216 +16,6 @@ interface SearchFieldProps extends Omit<InputHTMLAttributes<HTMLInputElement>, '
     variant?: FieldVariant
     assetType: AssetType
     onSelect?: (isin: string) => void
-}
-
-interface ISearchApiData {
-    id: string
-    assetType: AssetType
-    getUrl: (searchString: string) => string
-}
-
-// API response types for different services
-interface AlphaVantageMatch {
-    '1. symbol': string
-    '2. name': string
-    '3. type': string
-    '4. region': string
-    '5. marketOpen': string
-    '6. marketClose': string
-    '7. timezone': string
-    '8. currency': string
-    '9. matchScore': string
-}
-
-interface AlphaVantageResponse {
-    bestMatches: AlphaVantageMatch[]
-}
-
-interface CoinPaprikaResult {
-    id: string
-    name: string
-    symbol: string
-    rank: number
-    type: string
-}
-
-interface CoinPaprikaExchange {
-    id: string
-    name: string
-}
-
-interface CoinPaprikaIco {
-    id: string
-    name: string
-}
-
-interface CoinPaprikaPerson {
-    id: string
-    name: string
-}
-
-interface CoinPaprikaTag {
-    id: string
-    name: string
-}
-
-interface CoinPaprikaResponse {
-    currencies: CoinPaprikaResult[]
-    exchanges: CoinPaprikaExchange[]
-    icos: CoinPaprikaIco[]
-    people: CoinPaprikaPerson[]
-    tags: CoinPaprikaTag[]
-}
-
-type SearchResult = {
-    symbol: string
-    name: string
-    type?: string
-}
-
-const SEARCH_API_DATA: ISearchApiData[] = [
-    {
-        id: '1',
-        assetType: 'stock',
-        getUrl: (searchString: string) =>
-            `https://www.alphavantage.co/query?function=SYMBOL_SEARCH&keywords=${encodeURIComponent(
-                searchString,
-            )}&apikey=Y3E5063QA9MVU741`,
-    },
-    {
-        id: '2',
-        assetType: 'etf',
-        getUrl: (searchString: string) =>
-            `https://www.alphavantage.co/query?function=SYMBOL_SEARCH&keywords=${encodeURIComponent(
-                searchString,
-            )}&apikey=Y3E5063QA9MVU741`,
-    },
-    {
-        id: '3',
-        assetType: 'bond',
-        getUrl: (searchString: string) => `https://www.someFutureApi.co/${encodeURIComponent(searchString)}`,
-    },
-    {
-        id: '4',
-        assetType: 'crypto',
-        getUrl: (searchString: string) => `https://api.coinpaprika.com/v1/search?q=${encodeURIComponent(searchString)}`,
-    },
-]
-
-// Debounce hook
-function useDebounce<T>(value: T, delay: number): T {
-    const [debouncedValue, setDebouncedValue] = useState<T>(value)
-
-    useEffect(() => {
-        const handler = setTimeout(() => {
-            setDebouncedValue(value)
-        }, delay)
-
-        return () => {
-            clearTimeout(handler)
-        }
-    }, [value, delay])
-
-    return debouncedValue
-}
-
-// Fetch function
-const fetchSearchResults = async (searchString: string, assetType: AssetType): Promise<SearchResult[]> => {
-    if (!searchString.trim()) {
-        return []
-    }
-
-    const apiConfig = SEARCH_API_DATA.find((config) => config.assetType === assetType)
-    if (!apiConfig) {
-        throw new Error(`No API configuration found for asset type: ${assetType}`)
-    }
-
-    const url = apiConfig.getUrl(searchString)
-    const response = await fetch(url)
-
-    if (!response.ok) {
-        throw new Error(`Failed to search ${assetType}: ${response.statusText}`)
-    }
-
-    const data = await response.json()
-    console.log('data', data)
-    // TODO: apply fallback APIs functionality
-
-    // Transform response based on asset type
-    switch (assetType) {
-        case 'stock': {
-            const alphaData = data as AlphaVantageResponse
-            return (
-                alphaData.bestMatches
-                    ?.filter((match) => {
-                        const type = match['3. type'].toLowerCase()
-
-                        // Filtering for stocks only
-                        return (
-                            (type === 'equity' ||
-                                type.includes('common stock') ||
-                                type.includes('ordinary shares') ||
-                                type.includes('common share')) &&
-                            !type.includes('etf') &&
-                            !type.includes('fund') &&
-                            !type.includes('trust') &&
-                            !type.includes('reit') &&
-                            !type.includes('adr') &&
-                            !type.includes('warrant') &&
-                            !type.includes('preferred') &&
-                            !type.includes('bond')
-                        )
-                    })
-                    .map((match) => ({
-                        symbol: match['1. symbol'],
-                        name: match['2. name'],
-                        type: match['3. type'],
-                    })) || []
-            )
-        }
-
-        case 'etf': {
-            const alphaData = data as AlphaVantageResponse
-            return (
-                alphaData.bestMatches
-                    ?.filter((match) => {
-                        const type = match['3. type'].toLowerCase()
-
-                        // Filtering for ETFs only
-                        return (
-                            type.includes('etf') ||
-                            type.includes('exchange traded fund') ||
-                            type.includes('exchange-traded fund') ||
-                            type.includes('exchange traded') ||
-                            (type.includes('fund') &&
-                                (type.includes('index') || type.includes('sector') || type.includes('commodity')))
-                        )
-                    })
-                    .map((match) => ({
-                        symbol: match['1. symbol'],
-                        name: match['2. name'],
-                        type: match['3. type'],
-                    })) || []
-            )
-        }
-
-        case 'crypto': {
-            const coinData = data as CoinPaprikaResponse
-            return (
-                coinData.currencies?.map((currency) => ({
-                    symbol: currency.symbol,
-                    name: currency.name,
-                    type: currency.type,
-                })) || []
-            )
-        }
-
-        case 'bond':
-            return [] // TODO: add handling future APIs
-
-        default:
-            return data || []
-    }
 }
 
 export function SearchField({
@@ -262,7 +55,7 @@ export function SearchField({
     }
 
     const handleResultSelect = (result: SearchResult) => {
-        const selectedValue = result.symbol // or result.name
+        const selectedValue = result.name // or result.symbol
         setSearchString(selectedValue)
         setShowResults(false)
 
@@ -301,14 +94,14 @@ export function SearchField({
 
     useEffect(() => {
         console.log('searchResults: ', searchResults)
-    }, [searchResults])
+        console.log('error', error)
+    }, [searchResults, error])
 
     return (
         <div className="relative">
             <label>
                 <span className="input-label">{label}</span>
                 <div className="relative">
-                    {/* TODO: bring error from failed search  */}
                     <input
                         {...inputProps}
                         className={`h-9 input-basic ${error ? 'border-error focus:border-error' : styles}`}
@@ -337,19 +130,18 @@ export function SearchField({
                 </div>
             </label>
 
-            {/* Search Results Dropdown */}
             {showResults && searchResults && searchResults.length > 0 && (
-                <div className="absolute z-10 w-full mt-1 bg-white border border-gray-300 rounded-md shadow-lg max-h-60 overflow-y-auto">
+                <div className="absolute z-10 w-full mt-1 bg-searchBg border border-primaryLight rounded-md max-h-80 overflow-y-auto">
                     {searchResults.map((result, index) => (
                         <div
                             key={`${result.symbol}-${index}`}
-                            className="px-4 py-2 hover:bg-gray-100 cursor-pointer border-b border-gray-100 last:border-b-0"
+                            className="px-4 py-2 hover:bg-bg cursor-pointer border-b border-primaryLight last:border-b-0"
                             onMouseDown={(e) => e.preventDefault()} // Prevent input blur
                             onClick={() => handleResultSelect(result)}
                         >
-                            <div className="font-medium text-sm">{result.symbol}</div>
-                            <div className="text-xs text-gray-600 truncate">{result.name}</div>
-                            {result.type && <div className="text-xs text-gray-500">{result.type}</div>}
+                            <div className="font-medium text-sm text-white/90">{result.symbol}</div>
+                            {result.type && <div className="text-xs text-white/30 ">{result.type}</div>}
+                            <div className="text-xs text-white/70 truncate">{result.name}</div>
                         </div>
                     ))}
                 </div>
