+++ src/constants/colors.constants.ts
@@ -6,6 +6,7 @@ export const COLORS = {
     secondaryLight: '#FACC8C', //light-orange
     secondaryDark: '#FD4729', //dark-orange
     bg: '#0E0E12', //dark-grey
+    searchBg: '#32323b', //dark-grey
     sidebar: '#050807', //dark-dark-grey
     modalBg: '#050807', //dark-dark-grey
     error: '#AA193D', //dark-red
