+++ src/services/queryClient.ts
@@ -1,0 +1,10 @@
+import { QueryClient } from '@tanstack/react-query'
+
+export const queryClient = new QueryClient({
+    defaultOptions: {
+        queries: {
+            staleTime: 1 * 60 * 1000,
+            gcTime: 24 * 60 * 60 * 1000,
+        },
+    },
+})
